# a7_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'LBAW Infinity Auction Web API'
 description: 'Web Resources Specification (A7) for Infinity Auction'
servers:
 - url: https://lbaw2271.lbaw.fe.up.pt/
   description: Production server
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form.'
      description: 'Provide login form UI. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'
      responses:
        '200':
          description: 'Ok. Show Log-in UI [UI02].'
    post:
      operationId: R102
      summary: 'R102: Login Action.'
      description: 'Processes the login form submission. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:          # <!--- form field name
                  type: string
                password:    # <!--- form field name
                  type: string
              required:
                    - email
                    - password

      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to Home page.'
                  value: '/'
                302Error:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
...
  /register:
    get:
      operationId: R104
      summary: 'R104: Register Form.'
      description: 'Provide new user registration form UI. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'
      responses:
        '200':
          description: 'Ok. Show Sign-Up UI [UI04].'

    post:
      operationId: R105
      summary: 'R105: Register Action.'
      description: 'Processes the new user registration form submission. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                gender:
                  type: string
                  enum: ['M', 'F', 'NB', 'O']
                phone:
                  type: string
                email:
                  type: string
                  format: email
                birthdate:
                  type: string
                  format: date
                profile_picture:
                  type: string
                  format: binary
                  description: 'Can be jpeg, jpg, png and gif'
                address:
                  type: string
                password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
              required:
                    - name
                    - gender
                    - phone
                    - email
                    - birthdate
                    - profile_picture
                    - address
                    - password
                    - confirm_password

      responses:
        '302':
          description: 'Redirect after processing new user information.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful Registration. Redirect to user profile.'
                  value: '/'
                302Failure:
                  description: 'Failed Registration. Redirect to Register Form.'
                  value: '/register'
  /forgot-password:
    get:
      operationId: R106
      summary: 'R106: Forgot Password page UI.'
      description: 'Provide forgot password form UI. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'

      responses:
        '200':
          description: 'Ok. Show Forgot-Password UI.'

    post:
      operationId: R107
      summary: 'R107: Forgot Password Action.'
      description: 'Sends an email to reconfigure password. Access: PUB.'
      tags:
        - 'M01: Registration and authentication'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
              - email

      responses:
        '302':
          description: 'Redirect after processing user request.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfuly sent recover password email. Redirect to Forgot-Password page UI.'
                  value: '/forgot-password'
                302Failure:
                  description: 'Failed to setn recover password email. Redirect to Forgot-Password Page UI.'
                  value: '/forgot-password'
  /reset-password/{token}:
    get:
      operationId: R108
      summary: 'R108: Reset-Password page UI.'
      description: 'Provide Reset-Password form UI. Access: USR.'
      tags:
        - 'M01: Registration and authentication'
      
      parameters:
        - in: path
          name: token
          schema:
            type: string
          required: true

      responses:
        '200':
          description: 'Ok. Show Reset-Password UI.'

    post:
      operationId: R109
      summary: 'R109: Reset Password Action.'
      description: 'Resets user password. Access: USR.'
      tags:
        - 'M01: Registration and authentication'

      parameters:
        - in: path
          name: token
          schema:
            type: string
          required: true

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
                token:
                  type: string
              required:
              - email
              - password
              - confirm_password
              - token
      
      responses:
        '302':
          description: 'Redirect after processing user request.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfuly recovered password. Redirect to login page UI.'
                  value: '/login'
                302Failure:
                  description: 'Failed to recover password email. Redirect to Reset-Password Page UI.'
                  value: '/reset-password/{token}'

  # M02: Home page and Static pages
...
  /user/{id}:
    get:
      operationId: R401
      summary: 'R401: My profile UI.'
      description: 'Show signed user profile, including personal information, form to edit personal information, bidding activity, owned auctions and followed auctions. Access: USR.'
      tags:
        - 'M04: Users'

      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer 
            minimum: 1

      responses:
        '200':
          description: 'Ok. Show My Profile UI [UI14]. Edit profile section UI [UI18]. Bidding History section UI [UI20]. Followed auctions section UI [UI17]. Owned auctions section UI [UI09]'
        '403':
          description: 'Error user does not have permissions.'
    
    post:
      operationId: R402
      summary: 'R402: Edit profile.'
      description: 'Process new profile information. Access: USR, ADM.'
      tags:
        - 'M04: Users'

      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer 
            minimum: 1
      
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                  format: email
                birthdate:
                  type: string
                  format: date
                address:
                  type: string
                profile_image:
                  type: string
                  format: binary
                  description: 'Can be jpeg, jpg, png and gif'
                password:
                  type: string
                  format: password
                confirm_password:
                  type: string
                  format: password
              required:
                    - name
                    - phone
                    - email
                    - birthdate
                    - address
                    - profile_image

      responses:
        '302':
          description: 'Redirect after processing new user information.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful edit. Redirect to user profile.'
                  value: '/user/{id}'
                302Failure:
                  description: 'Failed to edit. Redirect to user Form.'
                  value: '/user/{id}'
        '403':
          description: 'Error user does not have permissions.'
        '500': 
          description: 'Invalid database parameters.'
...
  /users/{id}:
    get:
      operationId: R404
      summary: 'R404: View user profile.'
      description: 'This resource fetchs the user profile page including private and public data. Depending on the role of the user, the resulting page hides sensitive information. The authenticated user that owns the account can see all the information. The admin can see all the information of every user account. Every other role can only see the pubic information. In the case of the admin account, it shows the admin center, containing the admin private info, reported auctions and reported users. This last information can only be consulted by the ADM role. Access: PUB, USR, ADM'
      tags:
        - 'M04: Users'

      parameters:
        - in: path
          name: id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Ok. Show Profile UI [UI14]. Edit profile section UI [UI18] available for the USR and ADM. Bidding History section UI [UI20] available for the USR and ADM. Bidding auctions section UI available for the USR and ADM. Followed auctions section UI [UI17] available for the USR and ADM roles. Owned auctions section UI [UI09] available for the USR, ADM and PUB roles. Wishlist section UI available for the USR and ADM.  Won auctions section UI available for the USR and ADM. Admin center section UI available for the ADM. Reported users section UI available for the ADM. Reported auctions section UI available for the ADM'
  /api/users/delete/{id}:
    delete:
      operationId: R405
      summary: 'R405: Delete User.'
      description: 'Delete User profile, by deleting personal information stored in the database. Access: USR, ADM.'
      tags:
        - 'M04: Users'
      
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            minimum: 1
          required: true
      
      responses:
        '204':
          description: 'User profile successfully deleted.'
        '403':
          description: 'Can be the result of 3 operations: unauthorized access; user has running auctions; user has the highest bid on a running auction.' 
  /api/users/addReview:
    post:
      operationId: R406
      summary: 'R406: Add Review.'
      description: 'Add a five star rating review to another user. Access: USR'
      tags:
        - 'M04: Users'
      
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                rate:
                  type: integer
                  minimum: 1
                  maximum: 5
                user_id:
                  type: integer
              required:
                    - rate
                    - user_id
      
      responses:
        '200':
          description: 'Review added successfully!'
        '403':
          description: 'You do not have permissions to add this review!'
        '500':
          description: 'You can only review each seller once!'
  /api/user/follow_auction:
    post:
      operationId: R407
      summary: 'R407: Follow auction.'
      description: 'Proccesses request to follow auction. Access: USR'
      tags:
        - 'M04: Users'

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                auction_id:
                  type: integer
                  minimum: 1
                user_id:
                  type: integer
                  minimum: 1
              required:
                - auction_id
                - user_id
      
      responses:
        '200': 
          description: 'Successfully followed target auction.'
  /api/user/unfollow_auction:
    post:
      operationId: R408
      summary: 'R408: Unfollow auction.'
      description: 'Proccesses request to unfollow auction. Access: USR'
      tags:
        - 'M04: Users'

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                auction_id:
                  type: integer
                  minimum: 1
                user_id:
                  type: integer
                  minimum: 1
              required:
                - auction_id
                - user_id
      
      responses:
        '200':
          description: 'Successfully unfollowed target auction.'
  /api/user/follow_term:
    post:
      operationId: R409
      summary: 'R409: Add term to wishlist.'
      description: 'Proccesses request to follow term. Access: USR'
      tags:
        - 'M04: Users'
    
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                search_term:
                  type: string
              required:
                - search_term
      
      responses:
        '200':
          description: 'Successfully added term to wishlist.'
  /api/user/unfollow_term:
    post:
      operationId: R410
      summary: 'R410: Remove term from wishlist.'
      description: 'Proccesses request to unfollow term. Access: USR'
      tags:
        - 'M04: Users'
    
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                search_term:
                  type: string
              required:
                - search_term
      
      responses:
        '200':
          description: 'Successfully removed term from wishlist.'
  /api/user/follows_term:
    post:
      operationId: R411
      summary: 'R411: Remove term from wishlist.'
      description: 'Proccesses request to unfollow term. Access: USR'
      tags:
        - 'M04: Users'
    
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                search_term:
                  type: string
              required:
                - search_term
      
      responses:
        '200':
          description: 'Successfully removed term from wishlist.'
...
  /deposit:
    get:
      operationId: R413
      summary: 'R413: Paypal redirect.'
      description: 'Fetchs paypal API for deposit. Access: USR'
      tags:
        - 'M04: Users'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                deposit_value:
                  type: integer
                  minimum: 5
                  maximum: 999999999
              required:
                - deposit_value

      responses:
        '200':
          description: 'Ok. Show Paypal page.'
...
  /withdraw:
    get:
      operationId: R416
      summary: 'R416: Paypal redirect.'
      description: 'Fetchs paypal API for Withdraw. Access: USR'
      tags:
        - 'M04: Users'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                withdraw_value:
                  type: integer
                  minimum: 5
                  maximum: 999999999
              required:
                - withdraw_value

      responses:
        '200':
          description: 'Ok. Show Paypal page.'
...
  /auctions/{auction_id}:
    get:
      operationId: R501
      summary: 'R501: View Auction.'
      description: 'Show auction page. Access: PUB.'
      tags:
        - 'M05: Auctions'

      parameters:
        - in: path
          name: auction_id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Ok. Show auction page UI [UI10]'
  /sell:
    get:
      operationId: R502
      summary: 'R502: Create Auction page UI.'
      description: 'Fetch Create Auction Form. Access: USR.'
      tags:
        - 'M05: Auctions'

      responses:
        '200':
          description: 'Ok. Show Create Auction Form UI [UI11].'
    
    post:
      operationId: R503
      summary: 'R503: Create Auction action.'
      description: 'Processes the creation form submission. Access: USR.'
      tags:
        - 'M05: Auctions'

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                base_price:
                  type: integer
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                buy_now:
                  type: integer
                  nullable: true
                categories:
                  type: array
                  items:
                    type: boolean
              required:
                - title
                - description
                - images
                - base_price
                - start_date
                - end_date
                - buy_now
                - categories

      responses:
        '302':
          description: 'Redirect after creating auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfully created auction. Redirect to auction page.'
                  value: '/auction/{id}'
                302Failure:
                  description: 'Failed to create auction. Redirect create auction page.'
                  value: '/sell'
        '403':
          description: 'Error admins do not create auctions.'
        '500': 
          description: 'Invalid database parameters.'
  /auctions/edit/{id}:
    get:
      operationId: R504
      summary: 'R504: Edit Auction page.'
      description: 'Fetch form to edit the auction. Access: OWN, ADM'
      tags:
        - 'M05: Auctions'
      
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Show edit form UI [UI12].'
    
    post:
      operationId: R505
      summary: 'R505 Edit Auction Action.'
      description: 'Processes form in order to edit the auction. Access: OWN, ADM.'
      
      tags:
        - 'M05: Auctions'

      parameters:
        - in: path
          name: id
          schema:
            type: integer
            minimum: 1
          required: true

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                base_price:
                  type: integer
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                buy_now:
                  type: integer
                  nullable: true
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                    description: 'Can be jpeg, jpg, png and gif'
              required:
                - title
                - description
                - base_price
                - start_date
                - end_date
                - buy_now

      responses:
        '302':
          description: 'Redirect after editing auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfully edited auction. Redirect to auction page.'
                  value: '/auction/{id}'
                302Failure:
                  description: 'Failed to edit auction. Redirect edit Auction page.'
                  value: '/auction/edit/{id}'
        '403':
          description: 'Error user does not have permissions.'
  /auctions/cancel/{id}:
    post:
      operationId: R506
      summary: 'R506: Cancel Auction.'
      description: 'Enables Owner to set auction state to cancelled. Access: OWN, ADM.'
      tags:
        - 'M05: Auctions'

      parameters:
        - in: path
          name: id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '302':
          description: 'Redirect after cancelling the auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfully canceled. Redirect to auction page.'
                  value: '/'
                302Failure:
                  description: 'Failed to cancel. Redirect to auction page.'
                  value: '/auction/{id}'
        '403':
          description: 'Error user does not have permissions.'
  /api/auctions/getAllBids/{auction_id}:
    get:
      operationId: R507
      summary: 'R507: Auctions bids'
      description: 'List the bids of a certain auction. Access: PUB'
      tags:
        - 'M05: Auctions'

      parameters:
        - in: path
          name: auction_id
          schema:
            type: integer
            minimum: 1
          required: true
      
      responses:
        '200':
          description: 'Successfully fetched data.'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bid_id:
                      type: string
                      minimum: 1
                    date:
                      type: string
                      format: date-time
                    amount:
                      type: integer
                      minimum: 1
                    user_id:
                      type: integer
                      minimum: 1
                    auction_id:
                      type: integer
                      minimum: 1
                    username:
                      type: string
                  required:
                    - bid_id
                    - date
                    - amount
                    - user_id
                    - auction_id
                    - username
                example:
                  - bid_id: 100
                    date: '2021-07-11 00:00:00+00'
                    amount: 300
                    user_id: 4
                    auction_id: 9
                    username: Manel
                  - bid_id: 32
                    date: '2021-07-09 00:00:00+00'
                    amount: 200
                    user_id: 2
                    auction_id: 1
                    username: Maria
  /api/auctions:
    post:
      operationId: R508
      summary: 'R508: Bid on auction.'
      description: 'User bids on the auction. Access: BDR.'
      tags:
        - 'M05: Auctions'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  minimum: 1
                amount:
                  type: integer
                  minimum: 1
                auction_id:
                  type: integer
                  minimum: 1
              required:
                - user_id
                - value
                - auction_id
      
      responses:
        '302':
          description: 'Redirect after biding on the auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfull bid. Redirect to auction page.'
                  value: '/auction/{id}'
                302Failure:
                  description: 'Failed to make a bid. Redirect to auction page.'
                  value: '/auction/{id}'
        '403':
          description: 'Error user does not have permissions.'
...
  /api/image/{id}:
    delete:
      operationId: R511
      summary: 'R511: Delete auction image.'
      description: 'This api deletes an image that is associated with a specific auction. Access: OWN, ADM'
      tags:
        - 'M05: Auctions'
      
      parameters:
          - in: path
            name: auction_id
            schema:
              type: integer
              minimum: 1
            required: true
    
      responses:
        '200':
          description: 'Ok. Image was successfully deleted.'
        '403':
          description: 'Error, user does not have the required permissions.'
...
  /auctions/checkout/{id}:
    get:
      operationId: R513
      summary: 'R513: Auction checkout UI.'
      description: 'Provide auction checkout form UI Access: BDR.'
      
      parameters:
        - in: path
          name: auction_id
          schema:
            type: integer
            minimum: 1
          required: true
        
      responses:
        '200':
          description: 'Ok. Show checkout UI.'
...
  /users/report/{id}:
    get:
      operationId: R603
      summary: 'R603: Get report form.'
      description: 'Provide user report form UI. Access: USR.'
      tags:
        - 'M06: Reports'
      
      parameters:
        - in: path
          name: reported_user_id
          schema:
            type: integer
            minimum: 1
          required: true

      responses:
        '200':
          description: 'Ok. Show Report form UI.'
...